refactored YAML files to be server based and no more manifest constants in the code.
