+++ app/how-long/page.tsx
@@ -1,21 +1,9 @@
-import fs from "fs";
-import path from "path";
 import { Suspense } from "react";
 import { HowLongClient } from "./how-long-client";
-
-function readHowLongOptions(): string[] {
-  const content = fs.readFileSync(
-    path.join(process.cwd(), "yaml/HOWLONG.yaml"),
-    "utf-8"
-  );
-  return content
-    .split("\n")
-    .filter((l) => l.startsWith("- "))
-    .map((l) => l.slice(2).trim());
-}
+import { getHowLongOptions } from "@/lib/config";
 
 export default function HowLongPage() {
-  const options = readHowLongOptions();
+  const options = getHowLongOptions();
   return (
     <Suspense>
       <HowLongClient options={options} />
